# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "description": "Authenticate user with email and password. Returns access token and sets refresh token in HTTP-only cookie.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User credentials",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "example1": {
                  "summary": "Standard login",
                  "value": {
                    "email": "user@example.com",
                    "password": "Password123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Access token returned, refresh token set in cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": "user123",
                    "username": "john_doe",
                    "email": "john@example.com"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid credentials"
          },
          "401": {
            "description": "Unauthorized - Invalid email or password"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "summary": "User login",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "description": "Register a new user account. Returns access token and sets refresh token in HTTP-only cookie.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "User registration data",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              },
              "examples": {
                "example1": {
                  "summary": "New user registration",
                  "value": {
                    "email": "newuser@example.com",
                    "username": "new_user",
                    "password": "Password123!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful. Access token returned, refresh token set in cookie.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                  "user": {
                    "id": "user456",
                    "username": "new_user",
                    "email": "newuser@example.com"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Email already taken or invalid data"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "summary": "User registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "description": "Get a new access token using the refresh token from HTTP-only cookie. Requires valid refresh token in cookie.",
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Access token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponseDto"
                },
                "example": {
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or expired refresh token"
          },
          "403": {
            "description": "Forbidden - Access denied"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "security": [
          {
            "refresh_token": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "description": "Log out user by clearing refresh token from database and cookie. Requires valid access token.",
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                },
                "example": {
                  "success": true,
                  "message": "Logged out successfully"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid access token"
          },
          "429": {
            "description": "Too many requests - Rate limit exceeded"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "User logout",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Chat Backend API",
    "description": "API for chat application",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "jwt"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "User Password",
            "example": "strongpassword123",
            "minLength": 6
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "UserInfoDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID",
            "example": "user123"
          },
          "username": {
            "type": "string",
            "description": "Username",
            "example": "john_doe"
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john@example.com"
          }
        },
        "required": [
          "id",
          "username",
          "email"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token for API authorization",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "User information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfoDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Username (3-50 characters)",
            "example": "user111",
            "minLength": 3,
            "maxLength": 50
          },
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "description": "User Password",
            "example": "strongpassword123",
            "minLength": 6
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token for API authorization",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "user": {
            "description": "User information",
            "allOf": [
              {
                "$ref": "#/components/schemas/UserInfoDto"
              }
            ]
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "RefreshResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "New JWT access token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Logout operation success status",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Logout message",
            "example": "Logged out successfully"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    }
  }
}